218: use AddressModel for all kind of addresses
- reduces duplication of code
- put mixins and extension in one place
