+++ lib/features/campaigns/models/posters/poster_create_model.dart
@@ -1,20 +1,15 @@
 import 'dart:typed_data';
 
+import 'package:gruene_app/app/services/nominatim_service.dart';
 import 'package:maplibre_gl/maplibre_gl.dart';
 
 class PosterCreateModel {
-  final String street;
-  final String houseNumber;
-  final String zipCode;
-  final String city;
+  final AddressModel address;
   final LatLng location;
   final Uint8List? photo;
 
   const PosterCreateModel({
-    required this.street,
-    required this.houseNumber,
-    required this.zipCode,
-    required this.city,
+    required this.address,
     this.photo,
     required this.location,
   });
