+++ lib/features/campaigns/models/posters/poster_update_model.dart
@@ -1,13 +1,11 @@
 import 'dart:typed_data';
 
+import 'package:gruene_app/app/services/nominatim_service.dart';
 import 'package:gruene_app/features/campaigns/models/posters/poster_detail_model.dart';
 
 class PosterUpdateModel {
   final String id;
-  final String street;
-  final String housenumber;
-  final String city;
-  final String zipCode;
+  final AddressModel address;
   final PosterStatus status;
   final String comment;
   final Uint8List? newPhoto;
@@ -15,10 +13,7 @@ class PosterUpdateModel {
 
   PosterUpdateModel({
     required this.id,
-    required this.street,
-    required this.housenumber,
-    required this.zipCode,
-    required this.city,
+    required this.address,
     required this.status,
     required this.comment,
     required this.removePreviousPhotos,
