+++ lib/features/campaigns/screens/flyer_detail.dart
@@ -46,7 +46,7 @@ class FlyerDetail extends StatelessWidget {
           Container(
             padding: EdgeInsets.symmetric(vertical: 4),
             child: Text(
-              t.campaigns.posters.info_details,
+              t.campaigns.info_details,
               style: theme.textTheme.labelSmall?.apply(color: Colors.black, fontWeightDelta: 2),
             ),
           ),
