+++ lib/features/campaigns/screens/poster_add_screen.dart
@@ -6,7 +6,7 @@ import 'package:gruene_app/app/services/nominatim_service.dart';
 import 'package:gruene_app/app/theme/theme.dart';
 import 'package:gruene_app/features/campaigns/helper/media_helper.dart';
 import 'package:gruene_app/features/campaigns/models/posters/poster_create_model.dart';
-import 'package:gruene_app/features/campaigns/screens/doors_add_screen.dart';
+import 'package:gruene_app/features/campaigns/screens/screen_extensions.dart';
 import 'package:gruene_app/features/campaigns/widgets/create_address_widget.dart';
 import 'package:gruene_app/features/campaigns/widgets/save_cancel_on_create_widget.dart';
 import 'package:gruene_app/i18n/translations.g.dart';
@@ -23,7 +23,7 @@ class PosterAddScreen extends StatefulWidget {
   State<StatefulWidget> createState() => _PostersAddState();
 }
 
-class _PostersAddState extends State<PosterAddScreen> with AddressMixin {
+class _PostersAddState extends State<PosterAddScreen> with AddressExtension {
   @override
   TextEditingController streetTextController = TextEditingController();
   @override
@@ -143,10 +143,7 @@ class _PostersAddState extends State<PosterAddScreen> with AddressMixin {
       context,
       PosterCreateModel(
         location: widget.location,
-        street: streetTextController.text,
-        houseNumber: houseNumberTextController.text,
-        zipCode: zipCodeTextController.text,
-        city: cityTextController.text,
+        address: getAddress(),
         photo: reducedImage,
       ),
     );
