+++ lib/features/campaigns/screens/poster_detail.dart
@@ -33,7 +33,7 @@ class PosterDetailState extends State<PosterDetail> {
           ),
         ),
         Text(
-          t.campaigns.posters.info_details,
+          t.campaigns.info_details,
           style: theme.textTheme.labelSmall?.apply(color: Colors.black, fontWeightDelta: 2),
         ),
       ],
